Move yml-import to pages folder
